+++ cmd/moombox/main.go
@@ -43,7 +43,7 @@ import (
 )
 
 var (
-	version = "2.2.8"
+	version = "2.2.9"
 	commit  = ""
 )